Add `haserrormsg` boolean attribute to jobset API response
This attribute allows to know if an error occurred or not: when an
error occurs, errormsg is not an empty string. Note we can not use the
errormsg attribute because it can be arbitrarily long and is excluded
from the jobset API response.

(cherry picked from commit c8983ca07602eaa21d640c5b473f699f59ed97d7)